+++ R/as-nlists.R
@@ -11,6 +11,15 @@ as_nlists <- function(x, ...) {
   UseMethod("as_nlists")
 }
 
+#' @rdname as_nlist
+#' @export
+as.nlists <- function(x, ...) {
+  deprecate_soft("0.1.1", 
+                 what = "nlist::as.nlists()", 
+                 with = "nlist::as_nlists()")
+  UseMethod("as_nlists")
+}
+
 #' @describeIn as_nlists Coerce list to nlists
 #' @export
 as_nlists.list <- function(x, ...) {