+++ NEWS.md
@@ -1,4 +1,4 @@
-- Soft deprecated `as.nlist()` for `as_nlist()`.
+- Soft deprecated `as.nlist()` and `as.nlists()` for `as_nlist()` and `as.nlists()`.
 - Added `as_data_term.nlist()` and `as_data_term.nlists()`
 - Added `tidy.nlists()`
 - Added `set_pars()`.